Renamed initial.db to initial.jet.
Previous versions of Android (pre-2.3, according to the internet)
compress your asset files. They also have a max size for compressed
asset files. The current database exceeds that limit. The workaround is
that Android does not compress file extensions that are already
compressed, so the issue does not arise.

From now on, the initial.db should be named initial.jet.